refactor: decouple CLI from downloader
