+++ src/index.ts
@@ -1,0 +1,2 @@
+export { runDownloader } from "./core/downloader";
+export { run } from "./app";
