+++ test/lib/cancellation.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect, beforeEach, jest } from "bun:test";
-import { cancellationToken } from "../cancellation";
+import { cancellationToken } from "../../src/lib/cancellation";
 
 describe("cancellation", () => {
   beforeEach(() => {
